refactor: LottoConcat->LottoManager 로 변경 및 로또 generate하는 기능 추가
